Fix failing SFTP transfer when a duplicate packet arrives.
Because of a signed/unsigned error during subtraction and comparison
the code would incorrectly interpret some duplicate packets as being
too far outside of the negotiated sftp window. As a result SFTP halts
the side-effect transfer with a hard error forcing the client to
disconnect the RPC2 connection.

+++ lib-src/rpc2/rpc2-src/sftp3.c
@@ -289,9 +289,9 @@ int sftp_DataArrived(RPC2_PacketBuffer *pBuff, struct SFTP_Entry *sEntry)
 	return(0); 
     }
 
-    moffset = pBuff->Header.SeqNumber-sEntry->RecvLastContig;
+    moffset = (long)pBuff->Header.SeqNumber - (long)sEntry->RecvLastContig;
 
-    if (moffset > sEntry->WindowSize) {
+    if (moffset > (long)sEntry->WindowSize) {
 	BOGOSITY(sEntry, pBuff);
 	return(-1);
     }
